% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.table.taxo.R
\name{agg.table.taxo}
\alias{agg.table.taxo}
\title{Table aggregation according to taxonomy
Aggregates an ASV/OTU table according to a given taxonomic level}
\usage{
agg.table.taxo(tab, tax.lvl = "genus", tax.table)
}
\arguments{
\item{tab}{An abundance table containing samples in columns and OTUs/ASV in rows.}

\item{tax.lvl}{A string specifying the taxonomic level to be used for aggregation.
\code{tax.lvl} should be the name of one colum of tax.table (case insensitive).
The default is "genus".}

\item{tax.table}{A table containing the taxonomy of each ASV/OTU.}
}
\value{
An abundance table with the sum of counts for each ASV/OTU corresponding to
a given taxonomic level, found in each sample.
}
\description{
Table aggregation according to taxonomy
Aggregates an ASV/OTU table according to a given taxonomic level
}
