% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.blind.R
\name{rf.blind}
\alias{rf.blind}
\title{Blind cross validation for random forest}
\usage{
rf.blind(tab, treat, train.id = NA, mtry = NULL, n.tree = 500, n.forest = 100)
}
\arguments{
\item{tab}{An abundance table containing samples in columns and OTUs/ASV in rows.}

\item{treat}{A vector containing the class identity of each sample.}

\item{train.id}{A charecter sting to be searched in samples names that will be used for training.}

\item{mtry}{The mtry parameter to be passed to the \code{ranger} function.
See \code{ranger} documentation for details.}

\item{n.tree}{The number of tree to grow. The default is \code{500}.}

\item{n.forest}{The number of forests to grow. The default is \code{100}.}

\item{seed}{A number to set seed before sampling samples in the n-folding process
and before growing each forest. The default is \code{NA}.}
}
\value{
Returns a list object containing the confusion matrix, the error rate, the sensitivity
the precision as well as the variable importance obtained for each of the \code{n.forest} grown
forests.
}
\description{
Grows multiple random forests with blind cross validation: the algorithm is trained
a specific part of the dataset, and prediction are done on another part of the dataset.
}
