% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.blind.R
\name{rf.blind}
\alias{rf.blind}
\title{Non-random cross validation for random forest}
\usage{
rf.blind(
  tab,
  treat,
  train.id,
  mtry = NULL,
  n.tree = 500,
  n.forest = 10,
  seed = NULL
)
}
\arguments{
\item{tab}{An abundance or presence absence table containing samples in columns and OTUs/ASV in rows.}

\item{treat}{A boolean vector containing the class identity of each sample, i.e. the treatment to predict.
This means that you should pick a class as a reference for the calculation of precision and sensitivity.}

\item{train.id}{A charecter sting to be searched in samples names that will be used for training.
Can be a regular expression. Can alernatively be a boolean vector saying wether or not each sample
is part of the training dataset(TRUE for training samples, FALSE for testing samples), or a character
vector containing the training sample names.}

\item{mtry}{The mtry parameter to be passed to the \code{ranger} function.
See \code{ranger} documentation for details.}

\item{n.tree}{The number of tree to grow. The default is \code{500}.}

\item{n.forest}{The number of forests to grow. The default is \code{10}.}

\item{seed}{A number to set the seed before growing the forest. Only meaningful
if n.forest == 1. The default is \code{NULL}.}
}
\value{
A list object containing:
\itemize{
  \item a summary table with the number of true positives (TP), true negatives (TN), false positives (FP) and false negatives (FN)
the error rate, the sensistivity \eqn{TP/(TP + FN)}, and the precision \eqn{TP/(TP + FP)}
  \item The confusion matrix
  \item \code{n.forest} tables containing Gini index for each variable in each of the \code{n.forest} grown forests.
  This index gives the variable importance for classification.
}
}
\description{
Grows multiple random forests with non-random cross validation: the algorithm is trained on
a specific part of the dataset, and predictions are done on another part of the dataset.
}
