% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.kfold.R
\name{rf.kfold}
\alias{rf.kfold}
\title{k-fold cross validation for random forest}
\usage{
rf.kfold(tab, treat, k.fold = 5, mtry = NULL, n.tree = 500, seed = NULL)
}
\arguments{
\item{tab}{An abundance table containing samples in columns and OTUs/ASV in rows.}

\item{treat}{A vector containing the class identity of each sample.}

\item{k.fold}{A number of fold to be applied for k-fold cross-valisation.}

\item{mtry}{The mtry parameter to be passed to the \code{ranger} function. See \code{ranger} documentation for details.}

\item{n.tree}{The number of tree to grow in each of the \code{k} forests. The default is \code{500}.}

\item{seed}{A number to set the seed before before growing each forest. The default is \code{NULL}.}
}
\value{
A list object containing:
\itemize{
  \item a summary table with the number of true positives (TP), true negatives (TN), false positives (FP) and false negatives (FN)
the error rate, the sensistivity \eqn{TP/(TP + FN)}, and the precision \eqn{TP/(TP + FP)}
  \item The confusion matrix
  \item The a table containing Gini index for each variable. This index gives the variable importance for classification.
}
}
\description{
Splits the dataset in \code{k} and grows \code{k} random forests for classification, using alternatively each
of the \code{k} parts of the dataset to make predictions, while the other \code{k-1} parts are used for the training.
}
\examples{
tab <- asv_table_16S
z <- samples_16S
res <- rf.kfold(tab, treat=z$irrigation, k.fold = 5, mtry=540, seed = 1409)

}
