% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.opti.mtry.taxo.R
\name{rf.opti.mtry.taxo}
\alias{rf.opti.mtry.taxo}
\title{Random forest optimization
Runs random forest classification with several taxonomic level and mtry parameters and performs
k-fold or blind cross-validation}
\usage{
rf.opti.mtry.taxo(
  tab,
  tax.table,
  treat,
  n.mtry = 5,
  cross.val = "nfold",
  train.id = NA,
  n.tree = 100,
  rf.param = 5,
  seed = 1409
)
}
\description{
Random forest optimization
Runs random forest classification with several taxonomic level and mtry parameters and performs
k-fold or blind cross-validation
}
