% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.opti.mtry.taxo.R
\name{rf.opti.mtry.taxo}
\alias{rf.opti.mtry.taxo}
\title{Random forest optimisation}
\usage{
rf.opti.mtry.taxo(
  tab,
  tax.table,
  treat,
  n.mtry = 5,
  mtry = function(x) n * x/n.mtry,
  tax.lvl = c("ASV", "genus", "family", "order", "class"),
  cross.val = "kfold",
  train.id = NA,
  n.tree = 100,
  cross.param = 5,
  seed = 1409,
  RDSfile = NULL
)
}
\arguments{
\item{tab}{An abundance table containing samples in columns and OTUs/ASV in rows.}

\item{tax.table}{A table containing the taxonomy of each ASV/OTU.}

\item{treat}{A boolean vector containing the class identity of each sample, i.e. the treatment to predict.
This means that you should pick a class as a reference for the calculation of precision and sensitivity.}

\item{n.mtry}{The number of mtry parameters to be tested. Default is 5.}

\item{mtry}{A function of \code{x = ncol(tab)}, the number of variables (i.e. ASV or OTU)
and \code{n in 1:n.mtry}. Default is \code{function(x) n*x/n.mtry}.}

\item{tax.lvl}{A character vector containing the names of the taxonomic levels to be used for asv table
aggregation. Default is \code{c("ASV", "genus", "family", "order", "class")}.}

\item{cross.val}{The type of cross validation to perform. Possible values are "blind" or
"kfold" (Default).}

\item{train.id}{A string that matches the name of samples tu be used for training. Only
meaningful for \code{cross.val = "blind"}.}

\item{n.tree}{The number of tree to grow for each forest. Default is 100.}

\item{cross.param}{The parameter needed for cross validation: the number of folds for
\code{cross.val = "kfold"} or the number of forests to grow for \code{cross.val = "blind"}. Default is 5.}

\item{seed}{The seed to set before growing each forest, and before sampling of training dataset in
\code{cross.val = "kfold"}. Set to NA for no seeding. Default is \code{1409}.}

\item{RDSfile}{A string contaning the name of the RDS file to save the results.
Default is NULL and results are not saved.}
}
\value{
Returns a list of dataframes corresponding to the different taxonomic levels. Each dataframe contains
the confusion matrix, sensitivity, precision and error rate obtained for each value of the mtry parameter.
Mean value and standard deviation are computed over the results of the \code{cross.param} forests grown.
}
\description{
Runs random forest classification with several taxonomic level and mtry parameters and performs
k-fold or blind cross-validation.
}
